完成获取用户信息按学历划分接口

+++ src/main/java/com/catalystplus/admin/response/user/UserInfoResponse.java
@@ -18,9 +18,9 @@ public class UserInfoResponse implements Serializable {
     private String id;
 
     @ApiModelProperty("今日新增数量")
-    private Long addNumber;
+    private Integer addNumber;
 
     @ApiModelProperty("截至今日总数量")
-    private Long totalNumber;
+    private Integer totalNumber;
 
 }
